Add a Hilbert Transform operator (#72)
Add a Hilbert Transform operator, both in float -> complex
and stereo pair -> sse vector form. Add tests and update documentation.

Thanks to Sean Costello for providing the weights and idea to
use serial biquads.

+++ include/sst/basic-blocks/dsp/LanczosResampler.h
@@ -236,7 +236,7 @@ template <int bs> bool LanczosResampler<bs>::tablesInitialized{false};
 
 template <int bs> inline size_t LanczosResampler<bs>::populateNext(float *fL, float *fR, size_t max)
 {
-    int populated = 0;
+    size_t populated = 0;
     while (populated < max && (phaseI - phaseO) > A + 1)
     {
         read((phaseI - phaseO), fL[populated], fR[populated]);
